KM116: Do NOT attempt to find matching stacks if domain is empty string
... This will result in matching all stacks, and sending delete
to all of them.

+++ pkg/cognito/authcertdeleter.go
@@ -22,6 +22,10 @@ func (c *AuthCertDeleter) DeleteAuthCert(domain string) error {
 		return err
 	}
 
+	if domain == "" {
+		return nil
+	}
+
 	for _, stack := range stacklist.StackSummaries {
 		if strings.Contains(*stack.StackId, strings.ReplaceAll(domain, ".", "-")) {
 			_, err := c.usprovider.CloudFormation().DeleteStack(&cloudformation.DeleteStackInput{
